==> Script_05_compare_clocks.Rmd <==
---
title: "Markdown"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

#Script information

This script compares the 4 clocks in terms of how their CpGs overlap and correlate.

#Setup
```{r}
#Set working directory.
setwd("/home/tjonkman/researchdrive/tjonkman/epigenetic-clocks")

#set library path.
.libPaths("/home/tjonkman/researchdrive/tjonkman/epigenetic-clocks/Packages")

#Load all necessary libraries.
library(GenomicRanges)
library(Hmisc)
library(SummarizedExperiment)
library(gplots)

options(stringsAsFactors = F)
```

#Load data.
```{r}
load("Output/Output_02a_clocks.RData")
load("Input/Input_05a_heatmap_function.RData")
load("Output/Output_03a_twas_data.RData")
```

#Calculate the absolute number of overlapping CpGs between clocks.
```{r}
#hannum overlaps
OL11 <- length(na.omit(match(names(hannum), names(hannum))))
OL12 <- length(na.omit(match(names(hannum), names(horvath))))
OL13 <- length(na.omit(match(names(hannum), names(skinblood))))
OL14 <- length(na.omit(match(names(hannum), names(zhang))))

#horvath overlaps
OL21 <- length(na.omit(match(names(horvath), names(hannum))))
OL22 <- length(na.omit(match(names(horvath), names(horvath))))
OL23 <- length(na.omit(match(names(horvath), names(skinblood))))
OL24 <- length(na.omit(match(names(horvath), names(zhang))))

#skinblood overlaps
OL31 <- length(na.omit(match(names(skinblood), names(hannum))))
OL32 <- length(na.omit(match(names(skinblood), names(horvath))))
OL33 <- length(na.omit(match(names(skinblood), names(skinblood))))
OL34 <- length(na.omit(match(names(skinblood), names(zhang))))

#zhang overlaps
OL41 <- length(na.omit(match(names(zhang), names(hannum))))
OL42 <- length(na.omit(match(names(zhang), names(horvath))))
OL43 <- length(na.omit(match(names(zhang), names(skinblood))))
OL44 <- length(na.omit(match(names(zhang), names(zhang))))

clock.names <- c("Hannum Bld", "Horvath MT", "Horvath Skn/Bld", "Zhang Bld/Slv")
OLtab <- matrix(nrow = 5, ncol = 5, dimnames = list(c(clock.names, "Total CpGs"), c(clock.names, "Total CpGs")))
OLtab[1,1:4] <- as.numeric(c(OL11, OL12, OL13, OL14))
OLtab[2,1:4] <- as.numeric(c(OL21, OL22, OL23, OL24))
OLtab[3,1:4] <- as.numeric(c(OL31, OL32, OL33, OL34))
OLtab[4,1:4] <- as.numeric(c(OL41, OL42, OL43, OL44))

OLtab[5,] <- c(length(hannum), length(horvath), length(skinblood), length(zhang), length(unique(names(all.clocks))))
OLtab[,5] <- OLtab[5,]
OLtab

write.csv(OLtab, file = "Output/Output_05a_clock_overlaps.csv")
```

#Make a boxplot figure per clock, which contains the maximum correlations of each of that clock's CpGs with any CpG of one of the three other clocks (for instance, for each hannum CpG, find the horvath CpG with the largest correlation). This way, you can get an impression if any of the CpGs of the other clocks contain the same information as the CpGs of the index clock.
```{r}
# #Regress out age from the mvalues.
# fit <- lm(t(assay(mvalues.clocks)) ~ mvalues.clocks$sampling_age)
# res <- t(fit$residuals)

#Calculate the correlation matrix. Also make an absolute copy of it to test whether correlations are different from 0 in either direction. (change assay(mvalues) to res if you choose to regress out age)
cor.clocks <- cor(t(assay(mvalues.clocks)), method = "pearson")
cor.abs <- abs(cor.clocks)
cor.abs[1:10,1:10]

#Make a copy of the absolute correlations within a clock; for this, the correlations of 1 are filtered out.
cor.abs.internal <- cor.abs
cor.abs.internal[as.numeric(cor.abs.internal) == 1] <- 0
cor.abs.internal[1:10,1:10]

#Make a vector containing the maximum correlation of each clock-clock combination of CpGs (e.g. for each Hannum CpG, find the Horvath CpG which has the strongest correlation).
#NB: for correlations between 2 clocks, I chose to leave in the correlations of 1, which represent CpGs which are shared between the 2 clocks. However, for internal correlations, I filtered them out, since the 1's are meaningless here.
Find.Max.Correlations <- function(cors, set1, set2){
  cor.subset <- cors[rownames(cors) %in% names(set1),rownames(cors) %in% names(set2)]
  apply(cor.subset, 1, max)
}

cor.11 <- Find.Max.Correlations(cor.abs.internal, hannum, hannum)
cor.12 <- Find.Max.Correlations(cor.abs, hannum, horvath)
cor.13 <- Find.Max.Correlations(cor.abs, hannum, skinblood)
cor.14 <- Find.Max.Correlations(cor.abs, hannum, zhang)

cor.21 <- Find.Max.Correlations(cor.abs, horvath, hannum)
cor.22 <- Find.Max.Correlations(cor.abs.internal, horvath, horvath)
cor.23 <- Find.Max.Correlations(cor.abs, horvath, skinblood)
cor.24 <- Find.Max.Correlations(cor.abs, horvath, zhang)

cor.31 <- Find.Max.Correlations(cor.abs, skinblood, hannum)
cor.32 <- Find.Max.Correlations(cor.abs, skinblood, horvath)
cor.33 <- Find.Max.Correlations(cor.abs.internal, skinblood, skinblood)
cor.34 <- Find.Max.Correlations(cor.abs, skinblood, zhang)

cor.41 <- Find.Max.Correlations(cor.abs, zhang, hannum)
cor.42 <- Find.Max.Correlations(cor.abs, zhang, horvath)
cor.43 <- Find.Max.Correlations(cor.abs, zhang, skinblood)
cor.44 <- Find.Max.Correlations(cor.abs.internal, zhang, zhang)

round(rbind(summary(cor.11), summary(cor.12), summary(cor.13), summary(cor.14)), digits = 2)
round(rbind(summary(cor.21), summary(cor.22), summary(cor.23), summary(cor.24)), digits = 2)
round(rbind(summary(cor.31), summary(cor.32), summary(cor.33), summary(cor.34)), digits = 2)
round(rbind(summary(cor.41), summary(cor.42), summary(cor.43), summary(cor.44)), digits = 2)

#Merge the maximum correlations of each hannum CpG with all CpGs of each clock (including itself). Sort the dataframe by highest internal correlations.
library(reshape2)
cor.hannum <- as.data.frame(rbind(cor.11, cor.12, cor.13, cor.14))
cor.hannum <- cor.hannum[,order(cor.hannum[1,])]
cor.hannum$clock.1 <- clock.names[1]
cor.hannum$clock.2 <- clock.names
cor.hannum
melt.hannum <- melt(cor.hannum, id.vars = c("clock.1", "clock.2"))
colnames(melt.hannum) <- c("clock.1", "clock.2", "CpG", "max.cor")
head(melt.hannum)

#Merge the maximum correlations of each horvath CpG with all CpGs of each clock (including itself). Sort the dataframe by highest internal correlations.
cor.horvath <- as.data.frame(rbind(cor.21, cor.22, cor.23, cor.24))
cor.horvath <- cor.horvath[,order(cor.horvath[2,])]
cor.horvath$clock.1 <- clock.names[2]
cor.horvath$clock.2 <- clock.names
cor.horvath
melt.horvath <- melt(cor.horvath, id.vars = c("clock.1", "clock.2"))
colnames(melt.horvath) <- c("clock.1", "clock.2", "CpG", "max.cor")
head(melt.horvath)

#Merge the maximum correlations of each skinblood CpG with all CpGs of each clock (including itself). Sort the dataframe by highest internal correlations.
cor.skinblood <- as.data.frame(rbind(cor.31, cor.32, cor.33, cor.34))
cor.skinblood <- cor.skinblood[,order(cor.skinblood[3,])]
cor.skinblood$clock.1 <- clock.names[3]
cor.skinblood$clock.2 <- clock.names
cor.skinblood
melt.skinblood <- melt(cor.skinblood, id.vars = c("clock.1", "clock.2"))
colnames(melt.skinblood) <- c("clock.1", "clock.2", "CpG", "max.cor")
head(melt.skinblood)

#Merge the maximum correlations of each zhang CpG with all CpGs of each clock (including itself). Sort the dataframe by highest internal correlations.
cor.zhang <- as.data.frame(rbind(cor.41, cor.42, cor.43, cor.44))
cor.zhang <- cor.zhang[,order(cor.zhang[4,])]
cor.zhang$clock.1 <- clock.names[4]
cor.zhang$clock.2 <- clock.names
cor.zhang
melt.zhang <- melt(cor.zhang, id.vars = c("clock.1", "clock.2"))
colnames(melt.zhang) <- c("clock.1", "clock.2", "CpG", "max.cor")
head(melt.zhang)

#Make 4 different figures, each with the max correlations of a clock with the other clocks.
plot.theme <- list(
  theme_grey(base_size = 26), 
  theme(plot.title = element_blank(), 
        axis.title.x = element_blank(), axis.title.y = element_blank(), 
        axis.text.y=element_text(face="bold"), axis.text.x = element_text(size = 18, face = "bold", angle = 0, hjust = 0.5) 
        ),
  coord_cartesian(ylim = c(0, 1)),
  scale_y_continuous(breaks=c(0, 0.2, 0.4, 0.6, 0.8, 1)),
  ggtitle("Maximum correlation of clock CpGs"),
  xlab("Clock"),
  ylab("Maximum correlation")
)
contour.colors <- c("#ff3333", "#0066ff", "#00cc00", "#9900ff")

#hannum
test <- melt.hannum[melt.hannum$clock.1 != melt.hannum$clock.2,]
test$clock.2 <- as.factor(test$clock.2)
levels(test$clock.2) <- gsub(" ", "\n", levels(test$clock.2))
ggplot(data = test, aes(x = clock.2, y = max.cor, color = clock.2)) + geom_boxplot(width = 0.6, lwd = 2, outlier.size = 3, color = contour.colors[1]) + plot.theme
ggsave(filename = "Output/Plot_05a.1_Maximum_correlations_hannum.png", width = 5, height = 4, dpi = 300)

#horvath
test <- melt.horvath[melt.horvath$clock.1 != melt.horvath$clock.2,]
test$clock.2 <- as.factor(test$clock.2)
test$clock.2 <- factor(test$clock.2, levels = c(clock.names[2], clock.names[-2]))
levels(test$clock.2) <- gsub(" ", "\n", levels(test$clock.2))
ggplot(data = test, aes(x = clock.2, y = max.cor, color = clock.2)) + geom_boxplot(width = 0.6, lwd = 2, outlier.size = 3, color = contour.colors[2]) + plot.theme
ggsave(filename = "Output/Plot_05a.1_Maximum_correlations_horvath.png", width = 5, height = 4, dpi = 300)

#skinblood
test <- melt.skinblood[melt.skinblood$clock.1 != melt.skinblood$clock.2,]
test$clock.2 <- as.factor(test$clock.2)
test$clock.2 <- factor(test$clock.2, levels = c(clock.names[3], clock.names[-3]))
levels(test$clock.2) <- gsub(" ", "\n", levels(test$clock.2))
ggplot(data = test, aes(x = clock.2, y = max.cor, color = clock.2)) + geom_boxplot(width = 0.6, lwd = 2, outlier.size = 3, color = contour.colors[3]) + plot.theme
ggsave(filename = "Output/Plot_05a.1_Maximum_correlations_skinblood.png", width = 5, height = 4, dpi = 300)

#zhang
test <- melt.zhang[melt.zhang$clock.1 != melt.zhang$clock.2,]
test$clock.2 <- as.factor(test$clock.2)
test$clock.2 <- factor(test$clock.2, levels = c(clock.names[4], clock.names[-4]))
levels(test$clock.2) <- gsub(" ", "\n", levels(test$clock.2))
ggplot(data = test, aes(x = clock.2, y = max.cor, color = clock.2)) + geom_boxplot(width = 0.6, lwd = 2, outlier.size = 3, color = contour.colors[4]) + plot.theme
ggsave(filename = "Output/Plot_05a.1_Maximum_correlations_zhang.png", width = 5, height = 4, dpi = 300)
```

#Make a bar chart which shows, per clock pair, the number of CpGs which have a proxy with at least a correlation of 0.5, 0.7, or 0.9.
```{r}
proxys <- data.frame(
  Correlation = rep(c(">0.5", ">0.7", ">0.9"), 12),
  Index.clock = c(rep(clock.names[1], 9), rep(clock.names[2], 9), rep(clock.names[3], 9), rep(clock.names[4], 9)),
  Percentage = NaN,
  Comparison.clock = c(
    rep(clock.names[2], 3), rep(clock.names[3], 3), rep(clock.names[4], 3),
    rep(clock.names[1], 3), rep(clock.names[3], 3), rep(clock.names[4], 3),
    rep(clock.names[1], 3), rep(clock.names[2], 3), rep(clock.names[4], 3),
    rep(clock.names[1], 3), rep(clock.names[2], 3), rep(clock.names[3], 3))
)


#Fill in the values for hannum.
idx <- melt.hannum$max.cor[melt.hannum$clock.2 == clock.names[2]] > 0.5
proxys$Percentage[1] <- length(which(idx == T)) / length(idx)
idx <- melt.hannum$max.cor[melt.hannum$clock.2 == clock.names[2]] > 0.7
proxys$Percentage[2] <- length(which(idx == T)) / length(idx)
idx <- melt.hannum$max.cor[melt.hannum$clock.2 == clock.names[2]] > 0.9
proxys$Percentage[3] <- length(which(idx == T)) / length(idx)

idx <- melt.hannum$max.cor[melt.hannum$clock.2 == clock.names[3]] > 0.5
proxys$Percentage[4] <- length(which(idx == T)) / length(idx)
idx <- melt.hannum$max.cor[melt.hannum$clock.2 == clock.names[3]] > 0.7
proxys$Percentage[5] <- length(which(idx == T)) / length(idx)
idx <- melt.hannum$max.cor[melt.hannum$clock.2 == clock.names[3]] > 0.9
proxys$Percentage[6] <- length(which(idx == T)) / length(idx)

idx <- melt.hannum$max.cor[melt.hannum$clock.2 == clock.names[4]] > 0.5
proxys$Percentage[7] <- length(which(idx == T)) / length(idx)
idx <- melt.hannum$max.cor[melt.hannum$clock.2 == clock.names[4]] > 0.7
proxys$Percentage[8] <- length(which(idx == T)) / length(idx)
idx <- melt.hannum$max.cor[melt.hannum$clock.2 == clock.names[4]] > 0.9
proxys$Percentage[9] <- length(which(idx == T)) / length(idx)


#Fill in the values for horvath.
idx <- melt.horvath$max.cor[melt.horvath$clock.2 == clock.names[2]] > 0.5
proxys$Percentage[10] <- length(which(idx == T)) / length(idx)
idx <- melt.horvath$max.cor[melt.horvath$clock.2 == clock.names[2]] > 0.7
proxys$Percentage[11] <- length(which(idx == T)) / length(idx)
idx <- melt.horvath$max.cor[melt.horvath$clock.2 == clock.names[2]] > 0.9
proxys$Percentage[12] <- length(which(idx == T)) / length(idx)

idx <- melt.horvath$max.cor[melt.horvath$clock.2 == clock.names[3]] > 0.5
proxys$Percentage[13] <- length(which(idx == T)) / length(idx)
idx <- melt.horvath$max.cor[melt.horvath$clock.2 == clock.names[3]] > 0.7
proxys$Percentage[14] <- length(which(idx == T)) / length(idx)
idx <- melt.horvath$max.cor[melt.horvath$clock.2 == clock.names[3]] > 0.9
proxys$Percentage[15] <- length(which(idx == T)) / length(idx)

idx <- melt.horvath$max.cor[melt.horvath$clock.2 == clock.names[4]] > 0.5
proxys$Percentage[16] <- length(which(idx == T)) / length(idx)
idx <- melt.horvath$max.cor[melt.horvath$clock.2 == clock.names[4]] > 0.7
proxys$Percentage[17] <- length(which(idx == T)) / length(idx)
idx <- melt.horvath$max.cor[melt.horvath$clock.2 == clock.names[4]] > 0.9
proxys$Percentage[18] <- length(which(idx == T)) / length(idx)


#Fill in the values for skinblood.
idx <- melt.skinblood$max.cor[melt.skinblood$clock.2 == clock.names[2]] > 0.5
proxys$Percentage[19] <- length(which(idx == T)) / length(idx)
idx <- melt.skinblood$max.cor[melt.skinblood$clock.2 == clock.names[2]] > 0.7
proxys$Percentage[20] <- length(which(idx == T)) / length(idx)
idx <- melt.skinblood$max.cor[melt.skinblood$clock.2 == clock.names[2]] > 0.9
proxys$Percentage[21] <- length(which(idx == T)) / length(idx)

idx <- melt.skinblood$max.cor[melt.skinblood$clock.2 == clock.names[3]] > 0.5
proxys$Percentage[22] <- length(which(idx == T)) / length(idx)
idx <- melt.skinblood$max.cor[melt.skinblood$clock.2 == clock.names[3]] > 0.7
proxys$Percentage[23] <- length(which(idx == T)) / length(idx)
idx <- melt.skinblood$max.cor[melt.skinblood$clock.2 == clock.names[3]] > 0.9
proxys$Percentage[24] <- length(which(idx == T)) / length(idx)

idx <- melt.skinblood$max.cor[melt.skinblood$clock.2 == clock.names[4]] > 0.5
proxys$Percentage[25] <- length(which(idx == T)) / length(idx)
idx <- melt.skinblood$max.cor[melt.skinblood$clock.2 == clock.names[4]] > 0.7
proxys$Percentage[26] <- length(which(idx == T)) / length(idx)
idx <- melt.skinblood$max.cor[melt.skinblood$clock.2 == clock.names[4]] > 0.9
proxys$Percentage[27] <- length(which(idx == T)) / length(idx)


#Fill in the values for zhang.
idx <- melt.zhang$max.cor[melt.zhang$clock.2 == clock.names[2]] > 0.5
proxys$Percentage[28] <- length(which(idx == T)) / length(idx)
idx <- melt.zhang$max.cor[melt.zhang$clock.2 == clock.names[2]] > 0.7
proxys$Percentage[29] <- length(which(idx == T)) / length(idx)
idx <- melt.zhang$max.cor[melt.zhang$clock.2 == clock.names[2]] > 0.9
proxys$Percentage[30] <- length(which(idx == T)) / length(idx)

idx <- melt.zhang$max.cor[melt.zhang$clock.2 == clock.names[3]] > 0.5
proxys$Percentage[31] <- length(which(idx == T)) / length(idx)
idx <- melt.zhang$max.cor[melt.zhang$clock.2 == clock.names[3]] > 0.7
proxys$Percentage[32] <- length(which(idx == T)) / length(idx)
idx <- melt.zhang$max.cor[melt.zhang$clock.2 == clock.names[3]] > 0.9
proxys$Percentage[33] <- length(which(idx == T)) / length(idx)

idx <- melt.zhang$max.cor[melt.zhang$clock.2 == clock.names[4]] > 0.5
proxys$Percentage[34] <- length(which(idx == T)) / length(idx)
idx <- melt.zhang$max.cor[melt.zhang$clock.2 == clock.names[4]] > 0.7
proxys$Percentage[35] <- length(which(idx == T)) / length(idx)
idx <- melt.zhang$max.cor[melt.zhang$clock.2 == clock.names[4]] > 0.9
proxys$Percentage[36] <- length(which(idx == T)) / length(idx)

#For readibility, add a rounded percentage column.
proxys$Perc.100 <- round(proxys$Percentage, 2) * 100
proxys

#ggplot wants the percentages to exclude the other groups.
#For example, if 50% has a 0.5 proxy, 20% a 0.7 proxy, and 5% a 0.9 proxy, you need to mark the 0.5 group as 25% (50-20-5).
idx <-  1:nrow(proxys) %% 3
cor0.5 <- proxys[idx == 1,]
cor0.7 <- proxys[idx == 2,]
cor0.9 <- proxys[idx == 0,]

cor0.5$Percentage <- cor0.5$Percentage - cor0.7$Percentage
cor0.7$Percentage <- cor0.7$Percentage - cor0.9$Percentage

proxys[idx == 1,"Percentage"] <- cor0.5$Percentage
proxys[idx == 2,"Percentage"] <- cor0.7$Percentage
proxys

#Finally, prepare plot labels for proper formatting.
proxys$Comparison.clock <- gsub(pattern = " ", replacement = "\n", x = proxys$Comparison.clock)
proxys

#Plot the data.
ggplot(proxys, aes(fill=Correlation, y=Percentage, x=Comparison.clock)) + 
  theme_grey(base_size = 13) +
  theme(plot.title = element_blank(), 
        axis.text.x = element_text(face = "bold", hjust = 0.5), axis.text.y = element_text(face = "bold"),
        strip.text = element_text(size = 13),
        axis.title.x = element_blank(), axis.title.y = element_text(face = "bold"),
        legend.position = "top", legend.direction = "horizontal", legend.text = element_text()) +
  geom_bar(position="stack", stat="identity", color = "black", width = 0.6) +
  facet_wrap(~Index.clock, scales = "free", nrow = 2, ncol = 2) +
  scale_fill_brewer(palette = "Blues", guide = guide_legend(title.position = "top", label.position = "right", title.hjust = 0.5)) +
  scale_y_continuous(limits = c(0, 1), breaks = c(0, 0.2, 0.4, 0.6, 0.8, 1.0)) +
  ylab("Fraction")

ggsave(filename = "test/proxys.png", width = 6, height = 6)
```
